requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
google-genai
pillow
python-dotenv
gunicorn